moved files into group
