+++ app/(main)/home/file-upload-drawer.tsx
@@ -14,7 +14,7 @@ import CategorySelect from "./category-select";
 import SubmitDocumentButton from "@/components/submit-document-button";
 import { Input } from "@/components/ui/input";
 import { Textarea } from "@/components/ui/textarea";
-import { addNewDocumentAction } from "../actions";
+import { addNewDocumentAction } from "../../actions";
 import { useRef, useState } from "react";
 import { useToast } from "@/components/ui/use-toast";
 import { useQueryParam, withDefault, StringParam } from "use-query-params";
