+++ app/(main)/home/scan-button.tsx
