+++ app/(main)/layout.tsx
@@ -7,7 +7,9 @@ import { headers } from "next/headers";
 import { urbanist } from "@/lib/utils";
 
 import { Metadata } from "next";
-import ScanButton from "./scan-button";
+import ScanButton from "./home/scan-button";
+import SideMenu from "@/components/side-menu";
+import DesktopHeader from "@/components/desktop-header";
 
 export const metadata: Metadata = {
     title: 'Create Next App',
@@ -22,10 +24,18 @@ export default function HomeLayout({ children }: {
     const route = headers().get('next-url')!;
 
     return (
-        <main className={`${urbanist.className} pt-5 relative h-[100dvh]`}>
-            {children}
+        <div className={`${urbanist.className} max-sm:pt-5 relative h-[100dvh] md:grid min-h-screen w-full md:grid-cols-[220px_1fr] lg:grid-cols-[280px_1fr]`}>
+            <SideMenu/>
+            
+            <div>
+                <DesktopHeader/>
 
-            <footer className="w-full absolute bottom-0 grid grid-cols-3 bg-slate-300">
+                <main>{children}</main>
+            </div>
+
+
+            
+            <footer className="w-full md:hidden absolute bottom-0 grid grid-cols-3 bg-slate-300">
                 <Button variant={"ghost"}>
                     <BiSolidHomeAlt2 className={route?.includes('home') ? 'text-slate-900': 'text-white'}/>
                 </Button>
@@ -34,7 +44,7 @@ export default function HomeLayout({ children }: {
                     <FiInbox className={route?.includes('inbox') ? 'text-slate-900': 'text-white'}/>
                 </Button>
             </footer>
-        </main>
+        </div>
     )
 
 }
